V2V Comms using DSM
Applying Dynamic Spectrum Management (DSM) in V2V communications by selecting an appropriate frequency band through the selection of available licensed and unlicensed frequency bands for vehicles. Simulations done to identify the effect of interference in the dynamic spectrum by observing the path loss, SINR, and the throughput with various interfering users. The results show that the performance of the proposed DSM signals a significant improvement in the overall throughput and the signal-to-interference-plus-noise ratio (SINR) value is reduced by up to 60% compared to the fixed spectrum allocation. Although the dynamic spectrum is still affected by the interference from the existing cellular users, the throughput of the dynamic spectrum is still sufficient to transmit the information to other vehicles.